requirements.txt
streamlit==1.41.1
langchain==0.3.16
langchain-community==0.3.16
langchain-ollama==0.2.3
langchain-text-splitters==0.3.5
sentence-transformers==3.3.1
faiss-cpu==1.9.0
pdfplumber==0.11.5
pypdf==5.1.0
python-docx==1.1.2
rank-bm25==0.2.2
python-dotenv==1.0.1
pytest==8.3.4
